[verstion Final] finished up with sprites and window size

+++ include/mathHelper.c
@@ -6,7 +6,7 @@ float degToRad(int a)
     return ((a * M_PI) / 180.0);
 }
 
-int FixAng(int a)
+float FixAng(float a)
 { 
     if(a>359)
         a -= 360;
